nexus-device/requirements.txt
google-generativeai
speechrecognition
pyttsx3
python-docx
keyboard
pyaudio